update file index.jsp, script.js, style.css and application.properties
