+++ src/main/java/com/tinhocanhtrang/controller/HomeController.java
@@ -1,0 +1,14 @@
+package com.tinhocanhtrang.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+
+@Controller
+public class HomeController {
+	
+	@RequestMapping(value = "trangchu",method = RequestMethod.GET)
+	public String getHome() {
+		return "index";
+	}
+}
