+++ src/main/resources/application.properties
@@ -1,1 +1,3 @@
-
+server.port=8080
+spring.mvc.view.prefix=/views/
+spring.mvc.view.suffix=.jsp
